fix complex linting issues by splitting up MobileCoinClient to be more succint
